+++ src/tests/test_cases/sections.py
@@ -367,10 +367,7 @@ class TestGoogle:  # noqa: D203
         """
 
     @staticmethod
-    @expect("D417: Missing argument descriptions in the docstring "
-            "(argument(s) skip, verbose are missing descriptions in "
-            "'test_missing_docstring_another' docstring)", arg_count=2)
-    def test_missing_docstring_another(skip, verbose):  # noqa: D213, D407
+    def test_hanging_indent(skip, verbose):  # noqa: D213, D407
         """Do stuff.
 
         Args: