+++ docs/release_notes.rst
@@ -12,6 +12,10 @@ New Features
 
 * Add support for `property_decorators` config to ignore  D401
 
+Bug Fixes
+
+* Fix ``--match`` option to only consider filename when matching full paths (#550).
+
 6.1.1 - May 17th, 2021
 ---------------------------