+++ docs/release_notes.rst
@@ -12,6 +12,7 @@ New Features
 
 * Add support for `property_decorators` config to ignore D401.
 * Add support for Python 3.10 (#554).
+# Add reporting for private definitions (#562)
 
 6.1.1 - May 17th, 2021
 ---------------------------