+++ docs/release_notes.rst
@@ -10,8 +10,10 @@ Current Development Version
 
 New Features
 
+* Allow for hanging indent when documenting args in Google style. (#449)
 * Add support for `property_decorators` config to ignore D401.
 * Add support for Python 3.10 (#554).
+* Replace D10X errors with D419 if docstring exists but is empty (#559).
 
 Bug Fixes