+++ docs/release_notes.rst
@@ -18,6 +18,8 @@ New Features
 Bug Fixes
 
 * Fix ``--match`` option to only consider filename when matching full paths (#550).
+* Fix decorator parsing for async function. Resolves some false positives
+  with async functions and ``overload``. (#577)
 
 6.1.1 - May 17th, 2021
 ---------------------------