+++ docs/release_notes.rst
@@ -10,9 +10,15 @@ Current Development Version
 
 New Features
 
+* Allow for hanging indent when documenting args in Google style. (#449)
 * Add support for `property_decorators` config to ignore D401.
 * Add support for Python 3.10 (#554).
 * Change handling of "inaccessible" functions and classes (#561).
+* Replace D10X errors with D419 if docstring exists but is empty (#559).
+
+Bug Fixes
+
+* Fix ``--match`` option to only consider filename when matching full paths (#550).
 
 6.1.1 - May 17th, 2021
 ---------------------------